+++ src/Unified/ScreenUnifiedDashboard.cpp
@@ -5,6 +5,7 @@
 #include "InputEventPlus.h"
 #include "Economy/EconomyManager.h"
 #include "Network/ContentSwarmManager.h"
+#include "Network/SpectatorManager.h"
 #include "AssetSyncManager.h"
 #include "RageUtil.h"
 #include "GameSoundManager.h"
@@ -21,6 +22,7 @@ void ScreenUnifiedDashboard::Init()
 	m_Options.push_back( "WAGER BATTLE" ); // New
 	m_Options.push_back( "GYM CIRCUIT" );
 	m_Options.push_back( "TOURNAMENT LADDER" );
+	m_Options.push_back( "SPECTATE TV" ); // New
 	m_Options.push_back( "MARKETPLACE" ); // New
 	m_Options.push_back( "SYNC EXT. ASSETS" ); // New
 	m_Options.push_back( "CUSTOMIZE PROFILE" ); // New
@@ -137,6 +139,11 @@ void ScreenUnifiedDashboard::Input( const InputEventPlus &input )
 		else if( sChoice == "WAGER BATTLE" ) SCREENMAN->SetNewScreen( "ScreenBettingEntry" );
 		else if( sChoice == "GYM CIRCUIT" ) SCREENMAN->SetNewScreen( "ScreenGymWelcome" );
 		else if( sChoice == "TOURNAMENT LADDER" ) SCREENMAN->SetNewScreen( "ScreenTournamentLadder" );
+		else if( sChoice == "SPECTATE TV" )
+		{
+			SpectatorManager::Instance()->ConnectToMatch("Pro Finals: Chris vs FEFEMZ");
+			SCREENMAN->SystemMessage( "Connecting to Live Stream..." );
+		}
 		else if( sChoice == "MARKETPLACE" ) SCREENMAN->SetNewScreen( "ScreenMarketplace" );
 		else if( sChoice == "SYNC EXT. ASSETS" )
 		{